v2.11.00
Modified to allow for the SSH keys used by the application to be
supplied instead of auto-generated upon startup. Updated project
dependencies.

+++ src/main/java/com/keybox/common/db/DBInitServlet.java
@@ -15,6 +15,7 @@
  */
 package com.keybox.common.db;
 
+import com.keybox.common.util.AppConfig;
 import com.keybox.manage.model.Auth;
 import com.keybox.manage.util.DBUtils;
 import com.keybox.manage.util.EncryptionUtil;
@@ -24,6 +25,7 @@ import javax.servlet.ServletConfig;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import java.sql.Connection;
+import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.Statement;
 
@@ -46,22 +48,23 @@ public class DBInitServlet extends javax.servlet.http.HttpServlet {
 
         super.init(config);
 
-
+        //check if reset ssh application key is set
+        boolean resetSSHKey = "true".equals(AppConfig.getProperty("resetApplicationSSHKey"));
         try {
             Connection connection = DBUtils.getConn();
             Statement statement = connection.createStatement();
 
             ResultSet rs = statement.executeQuery("select * from information_schema.tables where upper(table_name) = 'USERS' and table_schema='PUBLIC'");
             if (rs == null || !rs.next()) {
+                resetSSHKey = true;
                 statement.executeUpdate("create table if not exists users (id INTEGER PRIMARY KEY AUTO_INCREMENT, first_nm varchar, last_nm varchar, email varchar, username varchar not null, password varchar, auth_token varchar, enabled boolean not null default true, user_type varchar not null default '" + Auth.ADMINISTRATOR + "')");
-                //insert default admin user
-                statement.executeUpdate("insert into users (username, password, user_type) values('admin', '" + EncryptionUtil.hash("changeme") + "', '" + Auth.MANAGER + "')");
+
 
                 statement.executeUpdate("create table if not exists system (id INTEGER PRIMARY KEY AUTO_INCREMENT, display_nm varchar not null, user varchar not null, host varchar not null, port INTEGER not null, authorized_keys varchar not null, status_cd varchar not null default 'INITIAL')");
                 statement.executeUpdate("create table if not exists profiles (id INTEGER PRIMARY KEY AUTO_INCREMENT, nm varchar not null, desc varchar not null)");
                 statement.executeUpdate("create table if not exists system_map (profile_id INTEGER, system_id INTEGER, foreign key (profile_id) references profiles(id) on delete cascade , foreign key (system_id) references system(id) on delete cascade, primary key (profile_id, system_id))");
                 statement.executeUpdate("create table if not exists user_map (user_id INTEGER, profile_id INTEGER, foreign key (user_id) references users(id) on delete cascade, foreign key (profile_id) references profiles(id) on delete cascade, primary key (user_id, profile_id))");
-                statement.executeUpdate("create table if not exists application_key (id INTEGER PRIMARY KEY AUTO_INCREMENT, public_key varchar not null, private_key varchar not null, passphrase varchar unique not null)");
+                statement.executeUpdate("create table if not exists application_key (id INTEGER PRIMARY KEY AUTO_INCREMENT, public_key varchar not null, private_key varchar not null, passphrase varchar)");
 
                 statement.executeUpdate("create table if not exists status (id INTEGER, user_id INTEGER, status_cd varchar not null default 'INITIAL', foreign key (id) references system(id) on delete cascade, foreign key (user_id) references users(id) on delete cascade)");
                 statement.executeUpdate("create table if not exists scripts (id INTEGER PRIMARY KEY AUTO_INCREMENT, user_id INTEGER, display_nm varchar not null, script varchar not null, foreign key (user_id) references users(id) on delete cascade)");
@@ -72,26 +75,53 @@ public class DBInitServlet extends javax.servlet.http.HttpServlet {
                 statement.executeUpdate("create table if not exists session_log (id BIGINT PRIMARY KEY AUTO_INCREMENT, user_id INTEGER, session_tm timestamp default CURRENT_TIMESTAMP, foreign key (user_id) references users(id) on delete cascade )");
                 statement.executeUpdate("create table if not exists terminal_log (session_id BIGINT, system_id INTEGER, output varchar not null, log_tm timestamp default CURRENT_TIMESTAMP, foreign key (session_id) references session_log(id) on delete cascade, foreign key (system_id) references system(id) on delete cascade)");
 
-                System.out.println("Generating KeyBox SSH public/private key pair");
-                //generate new key and insert passphrase
+                //insert default admin user
+                PreparedStatement pStmt = connection.prepareStatement("insert into users (username, password, user_type) values(?,?,?)");
+                pStmt.setString(1, "admin");
+                pStmt.setString(2, EncryptionUtil.hash("changeme"));
+                pStmt.setString(3, Auth.MANAGER);
+                pStmt.execute();
+                DBUtils.closeStmt(pStmt);
 
-                //generate application pub/pvt key and get vales
-                String passphrase=SSHUtil.keyGen();
-                String publicKey=SSHUtil.getPublicKey();
-                String privateKey=SSHUtil.getPrivateKey();
+            }
 
+            //if reset ssh application key then generate new key
+            if (resetSSHKey) {
                 //delete ssh keys
                 SSHUtil.deleteSshKeys();
 
-                statement.executeUpdate("insert into application_key (public_key, private_key, passphrase) values('"+publicKey+"', '" + EncryptionUtil.encrypt(privateKey)  + "','"+EncryptionUtil.encrypt(passphrase)+"')");
+                //delete old key entry
+                PreparedStatement pStmt = connection.prepareStatement("delete from application_key");
+                pStmt.execute();
+                DBUtils.closeStmt(pStmt);
+
+                //generate new key and insert passphrase
+                System.out.println("Setting KeyBox SSH public/private key pair");
+
+                //generate application pub/pvt key and get values
+                String passphrase = SSHUtil.keyGen();
+                String publicKey = SSHUtil.getPublicKey();
+                String privateKey = SSHUtil.getPrivateKey();
+
+                //insert new keys
+                pStmt = connection.prepareStatement("insert into application_key (public_key, private_key, passphrase) values(?,?,?)");
+                pStmt.setString(1, publicKey);
+                pStmt.setString(2, EncryptionUtil.encrypt(privateKey));
+                pStmt.setString(3, EncryptionUtil.encrypt(passphrase));
+                pStmt.execute();
+                DBUtils.closeStmt(pStmt);
+
                 System.out.println("KeyBox Public Key:");
                 System.out.println(publicKey);
 
-                passphrase=null;
-                publicKey=null;
-                privateKey=null;
+                passphrase = null;
+                publicKey = null;
+                privateKey = null;
 
+                //set to false
+                AppConfig.updateProperty("resetApplicationSSHKey", "false");
             }
+
 
             DBUtils.closeRs(rs);
             DBUtils.closeStmt(statement);
